# specs/001-saas-gate-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS Gate Management API
  version: 1.0.0
  description: |
    Multi-tenant gate access management system API.
    All tenant endpoints require JWT authentication with tenant context.
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users in tenant
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
    post:
      tags: [Users]
      summary: Create user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user details
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags: [Users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
    delete:
      tags: [Users]
      summary: Deactivate user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deactivated

  # ==================== GATES ====================
...
  /vehicles:
    get:
      tags: [Access]
      summary: List vehicles
      security:
        - bearerAuth: []
      parameters:
        - name: ownerId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vehicle list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
    post:
      tags: [Access]
      summary: Register vehicle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'
      responses:
        '201':
          description: Vehicle registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
  /vehicles/{vehicleId}:
    patch:
      tags: [Access]
      summary: Update vehicle
      security:
        - bearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleRequest'
      responses:
        '200':
          description: Vehicle updated
    delete:
      tags: [Access]
      summary: Delete vehicle
      security:
        - bearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vehicle deleted

  # ==================== ACCESS - VISITOR PASSES ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
          enum: [super_admin, building_admin, security, resident_owner, resident_tenant, staff_maid, staff_driver]
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        unitNumber:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, password, role, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        unitNumber:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        unitNumber:
          type: string
        isActive:
          type: boolean
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ===== TENANT =====
...
    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        rfidUid:
          type: string
        plateNumber:
          type: string
        make:
          type: string
        model:
          type: string
        color:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateVehicleRequest:
      type: object
      required: [plateNumber]
      properties:
        ownerId:
          type: string
          format: uuid
        rfidUid:
          type: string
        plateNumber:
          type: string
        make:
          type: string
        model:
          type: string
        color:
          type: string
    UpdateVehicleRequest:
      type: object
      properties:
        rfidUid:
          type: string
        plateNumber:
          type: string
        make:
          type: string
        model:
          type: string
        color:
          type: string
        isActive:
          type: boolean

    # ===== VISITOR PASS =====
...